Move extension's `setContext` to its own file
