+++ extensions/positron-r/src/commands.ts
@@ -1,17 +1,15 @@
 /*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2022 Posit Software, PBC. All rights reserved.
+ *  Copyright (C) 2023 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
 import * as positron from 'positron';
 import { delay } from './util';
 import { RRuntime } from './runtime';
+import { getRPackageName } from './contexts';
 import { randomUUID } from 'crypto';
 
 export async function registerCommands(context: vscode.ExtensionContext, runtimes: Map<string, RRuntime>) {
-
-	const isRPackage = await detectRPackage();
-	vscode.commands.executeCommand('setContext', 'isRPackage', isRPackage);
 
 	context.subscriptions.push(
 
@@ -140,36 +138,4 @@ export async function registerCommands(context: vscode.ExtensionContext, runtime
 			}
 		}),
 	);
-}
-
-async function detectRPackage(): Promise<boolean> {
-	const descriptionLines = await parseRPackageDescription();
-	const packageLines = descriptionLines.filter(line => line.startsWith('Package:'));
-	const typeLines = descriptionLines.filter(line => line.startsWith('Type:'));
-	const typeIsPackage = (typeLines.length > 0
-		? typeLines[0].toLowerCase().includes('package')
-		: false);
-	const typeIsPackageOrMissing = typeLines.length === 0 || typeIsPackage;
-	return packageLines.length > 0 && typeIsPackageOrMissing;
-}
-
-async function getRPackageName(): Promise<string> {
-	const descriptionLines = await parseRPackageDescription();
-	const packageLines = descriptionLines.filter(line => line.startsWith('Package:'))[0];
-	const packageName = packageLines.split(' ').slice(-1)[0];
-	return packageName;
-}
-
-async function parseRPackageDescription(): Promise<string[]> {
-	if (vscode.workspace.workspaceFolders !== undefined) {
-		const folderUri = vscode.workspace.workspaceFolders[0].uri;
-		const fileUri = vscode.Uri.joinPath(folderUri, 'DESCRIPTION');
-		try {
-			const bytes = await vscode.workspace.fs.readFile(fileUri);
-			const descriptionText = Buffer.from(bytes).toString('utf8');
-			const descriptionLines = descriptionText.split(/(\r?\n)/);
-			return descriptionLines;
-		} catch { }
-	}
-	return [''];
 }
